All buttons work as expected, dict from file
All conditions that could fail regarding buttons were fixed.
The process of sending data to Hardware is now on a thread, and progress bar working correctly. It can even pause the sending of data.
Currently testing the serial sending of the final data, the microcontroller must validate the string.
Next, check if required manual line break, and in microcontroller split strings that are 18 bits long into 3 chuncks.